repaces DTO's by use cases
