--- Core/Application/DataTransfertObjects/WishCreateRequest.cs
@@ -1,5 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Application.DataTransfertObjects;
-
-public record WishCreateRequest : WishUpdateRequest;
