--- Core/Application/DataTransfertObjects/WishResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Application.DataTransfertObjects;
-
-public record WishResponse : WishCreateRequest
-{
-    public string Id { get; init; }
-}
