+++ Core/Application/UseCases/Wishes/GetById/WishDetailedResponse.cs
@@ -1,0 +1,8 @@
+﻿using Application.DataTransfertObjects;
+
+namespace Application.UseCases.Wishes.GetById;
+
+public record WishDetailedResponse : WishResponse
+{
+    public string Id { get; init; }
+}
