+++ Core/Application/UseCases/Wishes/GetByQuery/WishResponse.cs
@@ -1,0 +1,8 @@
+﻿using Application.UseCases.Wishes.Create;
+
+namespace Application.DataTransfertObjects;
+
+public record WishResponse : WishCreateRequest
+{
+    public string Id { get; init; }
+}
