+++ Core/Application/UseCases/Wishes/Update/WishUpdateRequest.cs
@@ -1,0 +1,5 @@
+﻿using Application.UseCases.Wishes.Create;
+
+namespace Application.UseCases.Wishes.Update;
+
+public record WishUpdateRequest : WishCreateRequest;
