+++ Core/Application/UseCases/Wishes/Update/WishUpdateValidator.cs
@@ -1,28 +1,29 @@
-﻿using Application.DataTransfertObjects;
 using Domain.Repositories.Abstractions;
 using FluentValidation;
 
-namespace WebApi.Validators;
+namespace Application.UseCases.Wishes.Update;
 
-public class WishCreateValidator : AbstractValidator<WishCreateRequest>
+public class WishUpdateValidator : AbstractValidator<WishUpdateRequest>
 {
-    public WishCreateValidator(IWishesRepository wishesRepository)
+    public WishUpdateValidator(IWishesRepository wishesRepository)
     {
         RuleFor(wish => wish)
             .Must(HaveAtLeastOnePropertySet)
             .WithMessage("At least one of the wish properties must be set: Spiritually, FamiliallyRelationally, FinanciallyMaterially, ProfessionallyAcademically, Other");
             
+        /*
         RuleFor(wish => wish.Email)
             .NotEmpty()
             .MustAsync(async (wishId, cancellationToken) =>
             {
                 var wish = await wishesRepository.GetByIdAsync(wishId, cancellationToken);
-                return wish == null;
+                return wish != null;
             })
-            .WithMessage("Wish  with id '{PropertyValue}' already exist");
+            .WithMessage("Wish  with email '{PropertyValue}' doesn't already exist");
+        */
     }
 
-    private static bool HaveAtLeastOnePropertySet(WishCreateRequest request)
+    private static bool HaveAtLeastOnePropertySet(WishUpdateRequest request)
     {
         return !string.IsNullOrWhiteSpace(request.Spiritually) || !string.IsNullOrWhiteSpace(request.FamiliallyRelationally) || !string.IsNullOrWhiteSpace(request.FinanciallyMaterially) || !string.IsNullOrWhiteSpace(request.ProfessionallyAcademically) || !string.IsNullOrWhiteSpace(request.Other);
     }
